Make `StrimziVersionComparator` top-level class
